MVP product, looking decent

+++ Controllers/InvoicesController.cs
@@ -23,7 +23,8 @@ namespace EerCare.Controllers
         // GET: Invoices
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Invoice.ToListAsync());
+            var applicationDbContext = _context.Invoice.Include(i => i.Provider).Include(i => i.Member);
+            return View(await applicationDbContext.ToListAsync());
         }
 
         // GET: Invoices/Details/5
@@ -34,8 +35,9 @@ namespace EerCare.Controllers
                 return NotFound();
             }
 
-            var invoice = await _context.Invoice
+            var invoice = await _context.Invoice.Include(i => i.Provider).Include(i => i.Member).Include(i => i.InvoiceLineItems)
                 .FirstOrDefaultAsync(m => m.InvoiceId == id);
+
             if (invoice == null)
             {
                 return NotFound();
@@ -52,16 +54,16 @@ namespace EerCare.Controllers
 
             var invoice = new InvoiceViewModel();
 
-            invoice.allProviders = activeProviders.Select(p => new SelectListItem
+            invoice.allProviders = activeProviders.OrderBy(p => p.ProviderName).Select(p => new SelectListItem
             {
                 Value = p.ProviderId.ToString(),
                 Text = p.ProviderName
             }).ToList();
 
-            invoice.allMembers = activeMembers.Select(m => new SelectListItem
+            invoice.allMembers = activeMembers.OrderBy(m => m.AlphaMemberByLast).Select(m => new SelectListItem
             {
                 Value = m.Id.ToString(),
-                Text = m.LastName
+                Text = m.AlphaMemberByLast
             }).ToList();
 
             return View(invoice);
@@ -72,7 +74,7 @@ namespace EerCare.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("InvoiceId,ProviderId,MemberId,InvoiceDate,SettledDate")] Invoice invoice)
+        public async Task<IActionResult> Create([Bind("InvoiceId,ProviderId,MemberId,InvoiceDate")] Invoice invoice)
         {
             if (ModelState.IsValid)
             {
@@ -81,6 +83,33 @@ namespace EerCare.Controllers
                 return RedirectToAction(nameof(Index));
             }
             return View(invoice);
+        }
+
+        // GET: LineItems/Create
+        public async Task<IActionResult> NewLineItem(int id)
+        {
+            CreateLineItem lineItem = new CreateLineItem();
+            lineItem.InvoiceId = id;
+            ViewBag.returnUrl = Request.Headers["Referer"].ToString();
+
+            return View(lineItem);
+        }
+
+        // POST: LineItem/Create
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> NewLineItem([Bind("InvoiceId,LineItemId,ProcedureCode,ProcedureDesc,AmtBilled,AmtSettled")] string returnUrl, LineItem lineItem, int id)
+        {
+            if (ModelState.IsValid)
+            {
+                lineItem.InvoiceId = id;
+                _context.Add(lineItem);
+                await _context.SaveChangesAsync();
+                return Redirect(returnUrl);
+                //return RedirectToAction(nameof(Index));
+
+            }
+            return View(lineItem);
         }
 
         // GET: Invoices/Edit/5
@@ -92,11 +121,33 @@ namespace EerCare.Controllers
             }
 
             var invoice = await _context.Invoice.FindAsync(id);
+            
             if (invoice == null)
             {
                 return NotFound();
             }
-            return View(invoice);
+
+            var allProviders = await _context.Provider.ToListAsync();
+            var allMembers = await _context.Member.ToListAsync();
+
+            InvoiceViewModel editInvoice = new InvoiceViewModel();
+            {
+                editInvoice.Invoice = invoice;
+                editInvoice.allProviders = allProviders.OrderBy(p => p.ProviderName).Select(p => new SelectListItem
+                { 
+                    Value = p.ProviderId.ToString(),
+                    Text = p.ProviderName
+                }).ToList();
+
+                editInvoice.allMembers = allMembers.OrderBy(m => m.AlphaMemberByLast).Select(m => new SelectListItem
+                {
+                    Value = m.Id.ToString(),
+                    Text = m.AlphaMemberByLast
+                }).ToList();
+            }
+
+            ViewBag.returnUrl = Request.Headers["Referer"].ToString();
+            return View(editInvoice);
         }
 
         // POST: Invoices/Edit/5
@@ -104,7 +155,7 @@ namespace EerCare.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("InvoiceId,ProviderId,MemberId,InvoiceDate,SettledDate")] Invoice invoice)
+        public async Task<IActionResult> Edit(int id, [Bind("InvoiceId,ProviderId,MemberId,InvoiceDate,SettledDate")] Invoice invoice, string returnUrl)
         {
             if (id != invoice.InvoiceId)
             {
@@ -129,7 +180,7 @@ namespace EerCare.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                return Redirect(returnUrl);
             }
             return View(invoice);
         }
@@ -142,7 +193,7 @@ namespace EerCare.Controllers
                 return NotFound();
             }
 
-            var invoice = await _context.Invoice
+            var invoice = await _context.Invoice.Include(i => i.Provider).Include(i => i.Member).Include(i => i.InvoiceLineItems)
                 .FirstOrDefaultAsync(m => m.InvoiceId == id);
             if (invoice == null)
             {
@@ -157,7 +208,12 @@ namespace EerCare.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var invoice = await _context.Invoice.FindAsync(id);
+            var invoice = await _context.Invoice.Include(i => i.InvoiceLineItems).FirstOrDefaultAsync(m => m.InvoiceId == id);
+            if (invoice.InvoiceLineItems.Count > 0)
+            {
+                return View("DeleteErrorView");
+            }
+
             _context.Invoice.Remove(invoice);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
